marketplace_service/requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.5
djangorestframework==3.16.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_keras==0.1
termcolor==3.1.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.178
ultralytics-thop==2.0.15
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
